snapshot: add wrappedIterator, broken
