+++ counter.go
@@ -1,0 +1,6 @@
+package iterate
+
+type Counter interface {
+	Seen() int64
+	IsIndexing() bool
+}
